article_2/requirements.txt
langchain-huggingface==0.1.2
torch==2.6.0
loguru==0.7.3
chromadb==0.6.3
sentence-transformers==3.4.1
langchain-chroma==0.2.2
pydantic-settings==2.8.1
fastapi==0.115.12
uvicorn==0.34.0
aiosqlite==0.21.0
Jinja2==3.1.6
aiofiles==24.1.0
PyJWT==2.10.1
